refactor: move fifth edition template
